requirements.txt
python-telegram-bot==20.7
Flask==2.2.5
Pillow==9.5.0
gTTS==2.3.2
